+++ deliveries_server/src/main/java/tqsua/DeliveriesServer/controller/VehicleController.java
@@ -3,6 +3,7 @@ package tqsua.DeliveriesServer.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -13,6 +14,7 @@ import tqsua.DeliveriesServer.service.VehicleService;
 import javax.validation.Valid;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.HashMap;
 
 @RestController
 @RequestMapping("/api")
@@ -21,8 +23,10 @@ public class VehicleController {
     @Autowired
     private VehicleService vehicleService;
 
-    @GetMapping(path="/vehicles")
-    public ArrayList<VehicleDTO> getVehicles() throws IOException, InterruptedException {
+    @GetMapping(path="/private/vehicles")
+    public ArrayList<VehicleDTO> getVehicles(Authentication authentication) throws IOException, InterruptedException {
+        // TODO:
+        // Verify Admin
         ArrayList<VehicleDTO> vehicles = new ArrayList<>();
         for (Vehicle vehicle: this.vehicleService.getAllVehicles()) {
             vehicles.add(new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration()));
@@ -30,41 +34,95 @@ public class VehicleController {
         return vehicles;
     }
 
-    @GetMapping(path="/vehicle")
-    public VehicleDTO getVehicleById(@RequestParam(name="id") long id) {
+    @GetMapping(path="/private/vehicle")
+    public ResponseEntity<Object> getVehicleById(Authentication authentication, @RequestParam(name="id") long id) {
+        String rider_id = authentication.getName();
         Vehicle vehicle = this.vehicleService.getVehicleById(id);
-        if (vehicle==null) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        return new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration());
+
+        if (vehicle == null) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Not found.");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
+        if (!rider_id.equals(String.valueOf(vehicle.getRider().getId()))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
+        VehicleDTO response = new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration());
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @GetMapping(path="/vehiclesbyrider")
-    public ArrayList<VehicleDTO> getVehiclesByRiderId(@RequestParam(name="id") long id) {
+    @GetMapping(path="/private/rider/{id}/vehicles")
+    public ResponseEntity<Object> getVehiclesByRiderId(Authentication authentication ,@PathVariable(value="id") long id) {
+        String rider_id = authentication.getName();
+
+        if (!rider_id.equals(String.valueOf(id))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
+
         ArrayList<Vehicle> vehicleList = this.vehicleService.getVehiclesByRiderId(id);
         ArrayList<VehicleDTO> response = new ArrayList<>();
         for (Vehicle vehicle: vehicleList) {
             response.add(new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration()));
         }
-        return response;
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @PostMapping(path = "/vehicle")
+    @PostMapping(path = "/private/vehicle")
     @ResponseStatus(HttpStatus.CREATED)
-    public VehicleDTO createVehicle(@Valid @RequestBody VehicleDTO v) {
+    public ResponseEntity<Object> createVehicle(Authentication authentication ,@Valid @RequestBody VehicleDTO v) {
+        String id = authentication.getName();
+
+        if (!id.equals(String.valueOf(v.getRider()))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         Vehicle vehicle = this.vehicleService.saveVehicle(v);
         if (vehicle==null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        return new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration());
+        VehicleDTO response = new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration());
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
-    @PutMapping(path="/vehicle/{id}")
-    public ResponseEntity<VehicleDTO> updateVehicle(@PathVariable(value="id") Long id, @Valid @RequestBody VehicleDTO v) {
+    @PutMapping(path="/private/vehicle/{id}")
+    public ResponseEntity<Object> updateVehicle(Authentication authentication ,@PathVariable(value="id") Long id, @Valid @RequestBody VehicleDTO v) {
+        String rider_id = authentication.getName();
+        Vehicle vehicle_found = vehicleService.getVehicleById(id);
+        if (vehicle_found == null) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Not found");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
+
+        if (!rider_id.equals(String.valueOf(vehicle_found.getRider().getId()))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
+        
         Vehicle vehicle = this.vehicleService.updateVehicle(id, v);
-        if (vehicle==null) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         VehicleDTO updatedVehicle = new VehicleDTO(vehicle.getId(), vehicle.getBrand(), vehicle.getModel(), vehicle.getCategory(), vehicle.getCapacity(), vehicle.getRider().getId(), vehicle.getRegistration());
-        return ResponseEntity.ok(updatedVehicle);
+        return new ResponseEntity<>(updatedVehicle, HttpStatus.OK);
     }
 
-    @DeleteMapping(path="/vehicle/{id}")
-    public void deleteVehicle(@PathVariable(value="id") Long id) {
+    @DeleteMapping(path="/private/vehicle/{id}")
+    public ResponseEntity<Object> deleteVehicle(Authentication authentication ,@PathVariable(value="id") Long id) {
+        String rider_id = authentication.getName();
+        Vehicle v = this.vehicleService.getVehicleById(id);
+        if (v == null) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Not found.");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
+        if (!rider_id.equals(String.valueOf(v.getRider().getId()))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put("message", "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         if (this.vehicleService.deleteVehicle(id)==null) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
